Extract command initialization into partial class files
Phase 1 of ViewModel extraction - split InitializeCommands (~2371 lines)
into 6 partial class files for better organization:

- MainViewModel.Commands.Navigation.cs - Panel toggles, view modes, camera
- MainViewModel.Commands.Simulator.cs - GPS simulation controls
- MainViewModel.Commands.Configuration.cs - Data IO, profiles, config dialogs
- MainViewModel.Commands.Fields.cs - Field selection, creation, import
- MainViewModel.Commands.Boundary.cs - Boundary map, AgShare, headland
- MainViewModel.Commands.Track.cs - Track management, section control, guidance

MainViewModel.cs reduced from ~8438 to ~6071 lines.
