+++ Shared/AgValoniaGPS.ViewModels/MainViewModel.Commands.Navigation.cs
@@ -1,0 +1,102 @@
+using CommunityToolkit.Mvvm.Input;
+
+namespace AgValoniaGPS.ViewModels;
+
+/// <summary>
+/// Navigation panel commands - view toggles, camera controls, brightness.
+/// </summary>
+public partial class MainViewModel
+{
+    private void InitializeNavigationCommands()
+    {
+        // Panel toggle commands
+        ToggleViewSettingsPanelCommand = new RelayCommand(() =>
+        {
+            IsViewSettingsPanelVisible = !IsViewSettingsPanelVisible;
+        });
+
+        ToggleFileMenuPanelCommand = new RelayCommand(() =>
+        {
+            IsFileMenuPanelVisible = !IsFileMenuPanelVisible;
+        });
+
+        ToggleToolsPanelCommand = new RelayCommand(() =>
+        {
+            IsToolsPanelVisible = !IsToolsPanelVisible;
+        });
+
+        ToggleConfigurationPanelCommand = new RelayCommand(() =>
+        {
+            IsConfigurationPanelVisible = !IsConfigurationPanelVisible;
+        });
+
+        ToggleJobMenuPanelCommand = new RelayCommand(() =>
+        {
+            IsJobMenuPanelVisible = !IsJobMenuPanelVisible;
+        });
+
+        ToggleFieldToolsPanelCommand = new RelayCommand(() =>
+        {
+            IsFieldToolsPanelVisible = !IsFieldToolsPanelVisible;
+        });
+
+        // View mode commands
+        ToggleGridCommand = new RelayCommand(() =>
+        {
+            IsGridOn = !IsGridOn;
+        });
+
+        ToggleDayNightCommand = new RelayCommand(() =>
+        {
+            IsDayMode = !IsDayMode;
+        });
+
+        Toggle2D3DCommand = new RelayCommand(() =>
+        {
+            Is2DMode = !Is2DMode;
+        });
+
+        ToggleNorthUpCommand = new RelayCommand(() =>
+        {
+            IsNorthUp = !IsNorthUp;
+        });
+
+        // Camera controls
+        IncreaseCameraPitchCommand = new RelayCommand(() =>
+        {
+            CameraPitch += 5.0;
+        });
+
+        DecreaseCameraPitchCommand = new RelayCommand(() =>
+        {
+            CameraPitch -= 5.0;
+        });
+
+        // Brightness controls
+        IncreaseBrightnessCommand = new RelayCommand(() =>
+        {
+            Brightness += 5;
+        });
+
+        DecreaseBrightnessCommand = new RelayCommand(() =>
+        {
+            Brightness -= 5;
+        });
+
+        // iOS Sheet toggle commands
+        ToggleFileMenuCommand = new RelayCommand(() =>
+        {
+            IsFileMenuVisible = !IsFileMenuVisible;
+        });
+
+        ToggleFieldToolsCommand = new RelayCommand(() =>
+        {
+            IsFieldToolsVisible = !IsFieldToolsVisible;
+        });
+
+        ToggleSettingsCommand = new RelayCommand(() =>
+        {
+            IsSettingsVisible = !IsSettingsVisible;
+        });
+    }
+}
