+++ Shared/AgValoniaGPS.ViewModels/MainViewModel.cs
@@ -287,8 +287,14 @@ public partial class MainViewModel : ObservableObject
         };
         _simulatorTimer.Tick += OnSimulatorTick;
 
-        // Initialize commands immediately (with MainThreadScheduler they're thread-safe)
-        InitializeCommands();
+        // Initialize commands (split into partial class files for organization)
+        InitializeNavigationCommands();
+        InitializeSimulatorCommands();
+        InitializeConfigurationCommands();
+        InitializeFieldCommands();
+        InitializeBoundaryCommands();
+        InitializeTrackCommands();
+        InitializeNtripCommands();
 
         // Load display settings first, then restore our app settings on top
         // This ensures AppSettings takes precedence over DisplaySettings
@@ -4474,2378 +4480,6 @@ public partial class MainViewModel : ObservableObject
     public ICommand? ToggleYouTurnCommand { get; private set; }
     public ICommand? ToggleAutoSteerCommand { get; private set; }
 
-    private void InitializeCommands()
-    {
-        // Use simple RelayCommand to avoid ReactiveCommand threading issues
-        // Use property setters instead of service methods to ensure PropertyChanged fires
-        ToggleViewSettingsPanelCommand = new RelayCommand(() =>
-        {
-            IsViewSettingsPanelVisible = !IsViewSettingsPanelVisible;
-        });
-
-        ToggleFileMenuPanelCommand = new RelayCommand(() =>
-        {
-            IsFileMenuPanelVisible = !IsFileMenuPanelVisible;
-        });
-
-        ToggleToolsPanelCommand = new RelayCommand(() =>
-        {
-            IsToolsPanelVisible = !IsToolsPanelVisible;
-        });
-
-        ToggleConfigurationPanelCommand = new RelayCommand(() =>
-        {
-            IsConfigurationPanelVisible = !IsConfigurationPanelVisible;
-        });
-
-        ToggleJobMenuPanelCommand = new RelayCommand(() =>
-        {
-            IsJobMenuPanelVisible = !IsJobMenuPanelVisible;
-        });
-
-        ToggleFieldToolsPanelCommand = new RelayCommand(() =>
-        {
-            IsFieldToolsPanelVisible = !IsFieldToolsPanelVisible;
-        });
-
-        ToggleGridCommand = new RelayCommand(() =>
-        {
-            IsGridOn = !IsGridOn;
-        });
-
-        ToggleDayNightCommand = new RelayCommand(() =>
-        {
-            IsDayMode = !IsDayMode;
-        });
-
-        Toggle2D3DCommand = new RelayCommand(() =>
-        {
-            Is2DMode = !Is2DMode;
-        });
-
-        ToggleNorthUpCommand = new RelayCommand(() =>
-        {
-            IsNorthUp = !IsNorthUp;
-        });
-
-        IncreaseCameraPitchCommand = new RelayCommand(() =>
-        {
-            CameraPitch += 5.0;
-        });
-
-        DecreaseCameraPitchCommand = new RelayCommand(() =>
-        {
-            CameraPitch -= 5.0;
-        });
-
-        IncreaseBrightnessCommand = new RelayCommand(() =>
-        {
-            Brightness += 5; // Match the step from DisplaySettingsService
-        });
-
-        DecreaseBrightnessCommand = new RelayCommand(() =>
-        {
-            Brightness -= 5;
-        });
-
-        // iOS Sheet toggle commands
-        ToggleFileMenuCommand = new RelayCommand(() =>
-        {
-            IsFileMenuVisible = !IsFileMenuVisible;
-        });
-
-        ToggleFieldToolsCommand = new RelayCommand(() =>
-        {
-            IsFieldToolsVisible = !IsFieldToolsVisible;
-        });
-
-        ToggleSettingsCommand = new RelayCommand(() =>
-        {
-            IsSettingsVisible = !IsSettingsVisible;
-        });
-
-        // Simulator commands
-        ToggleSimulatorPanelCommand = new RelayCommand(() =>
-        {
-            IsSimulatorPanelVisible = !IsSimulatorPanelVisible;
-        });
-
-        ResetSimulatorCommand = new RelayCommand(() =>
-        {
-            _simulatorService.Reset();
-            SimulatorSteerAngle = 0;
-            StatusMessage = "Simulator Reset";
-        });
-
-        ResetSteerAngleCommand = new RelayCommand(() =>
-        {
-            SimulatorSteerAngle = 0;
-            StatusMessage = "Steer Angle Reset to 0°";
-        });
-
-        SimulatorForwardCommand = new RelayCommand(() =>
-        {
-            _simulatorService.StepDistance = 0;  // Reset speed before accelerating
-            _simulatorService.IsAcceleratingForward = true;
-            _simulatorService.IsAcceleratingBackward = false;
-            StatusMessage = "Sim: Accelerating Forward";
-        });
-
-        SimulatorStopCommand = new RelayCommand(() =>
-        {
-            _simulatorService.IsAcceleratingForward = false;
-            _simulatorService.IsAcceleratingBackward = false;
-            _simulatorService.StepDistance = 0;  // Immediately stop movement
-            _simulatorSpeedKph = 0;  // Reset speed slider (use backing field to avoid triggering setter again)
-            OnPropertyChanged(nameof(SimulatorSpeedKph));
-            OnPropertyChanged(nameof(SimulatorSpeedDisplay));
-            StatusMessage = "Sim: Stopped";
-        });
-
-        SimulatorReverseCommand = new RelayCommand(() =>
-        {
-            _simulatorService.StepDistance = 0;  // Reset speed before accelerating
-            _simulatorService.IsAcceleratingBackward = true;
-            _simulatorService.IsAcceleratingForward = false;
-            StatusMessage = "Sim: Accelerating Reverse";
-        });
-
-        SimulatorReverseDirectionCommand = new RelayCommand(() =>
-        {
-            // Reverse direction by adding 180 degrees to current heading
-            var newHeading = _simulatorService.HeadingRadians + Math.PI;
-            // Normalize to 0-2π range
-            if (newHeading > Math.PI * 2)
-                newHeading -= Math.PI * 2;
-            _simulatorService.SetHeading(newHeading);
-            StatusMessage = "Sim: Direction Reversed";
-        });
-
-        SimulatorSteerLeftCommand = new RelayCommand(() =>
-        {
-            SimulatorSteerAngle -= 5.0; // 5 degree increments
-            StatusMessage = $"Steer: {SimulatorSteerAngle:F1}°";
-        });
-
-        SimulatorSteerRightCommand = new RelayCommand(() =>
-        {
-            SimulatorSteerAngle += 5.0; // 5 degree increments
-            StatusMessage = $"Steer: {SimulatorSteerAngle:F1}°";
-        });
-
-        // Dialog Commands
-        ShowDataIODialogCommand = new RelayCommand(() =>
-        {
-            State.UI.ShowDialog(DialogType.DataIO);
-        });
-
-        CloseDataIODialogCommand = new RelayCommand(CloseDataIODialog);
-
-        // Configuration Dialog Commands
-        ShowConfigurationDialogCommand = new RelayCommand(() =>
-        {
-            ConfigurationViewModel = new ConfigurationViewModel(_configurationService);
-            ConfigurationViewModel.CloseRequested += (s, e) =>
-            {
-                ConfigurationViewModel.IsDialogVisible = false;
-            };
-            ConfigurationViewModel.IsDialogVisible = true;
-        });
-
-        CancelConfigurationDialogCommand = new RelayCommand(() =>
-        {
-            if (ConfigurationViewModel != null)
-                ConfigurationViewModel.IsDialogVisible = false;
-        });
-
-        ShowLoadProfileDialogCommand = new RelayCommand(() =>
-        {
-            // Refresh available profiles
-            AvailableProfiles.Clear();
-            foreach (var profile in _configurationService.GetAvailableProfiles())
-            {
-                AvailableProfiles.Add(profile);
-            }
-            SelectedProfile = _configurationService.Store.ActiveProfileName;
-            IsProfileSelectionVisible = true;
-        });
-
-        LoadSelectedProfileCommand = new RelayCommand(() =>
-        {
-            if (!string.IsNullOrEmpty(SelectedProfile))
-            {
-                _configurationService.LoadProfile(SelectedProfile);
-                _settingsService.Settings.LastUsedVehicleProfile = SelectedProfile;
-                _settingsService.Save();
-                OnPropertyChanged(nameof(CurrentProfileName));
-            }
-            IsProfileSelectionVisible = false;
-        });
-
-        CancelProfileSelectionCommand = new RelayCommand(() =>
-        {
-            IsProfileSelectionVisible = false;
-        });
-
-        ShowNewProfileDialogCommand = new RelayCommand(() =>
-        {
-            // Generate a unique profile name
-            var baseName = "New Profile";
-            var profileName = baseName;
-            var counter = 1;
-            var existingProfiles = _configurationService.GetAvailableProfiles();
-            while (existingProfiles.Contains(profileName))
-            {
-                profileName = $"{baseName} {counter++}";
-            }
-
-            _configurationService.CreateProfile(profileName);
-            _settingsService.Settings.LastUsedVehicleProfile = profileName;
-            _settingsService.Save();
-            OnPropertyChanged(nameof(CurrentProfileName));
-        });
-
-        ShowSimCoordsDialogCommand = new RelayCommand(() =>
-        {
-            if (IsSimulatorEnabled)
-            {
-                // Don't allow changing coords while simulator is running
-                System.Diagnostics.Debug.WriteLine("[SimCoords] Disable simulator first to change coordinates");
-                return;
-            }
-            // Load current position into the dialog fields
-            // Round to 8 decimal places
-            var currentPos = GetSimulatorPosition();
-            SimCoordsDialogLatitude = Math.Round((decimal)currentPos.Latitude, 8);
-            SimCoordsDialogLongitude = Math.Round((decimal)currentPos.Longitude, 8);
-            // Show the panel-based dialog
-            State.UI.ShowDialog(DialogType.SimCoords);
-        });
-
-        CancelSimCoordsDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-        });
-
-        ConfirmSimCoordsDialogCommand = new RelayCommand(() =>
-        {
-            // Apply the coordinates from the dialog (convert from decimal? to double)
-            double lat = (double)(SimCoordsDialogLatitude ?? 0m);
-            double lon = (double)(SimCoordsDialogLongitude ?? 0m);
-            SetSimulatorCoordinates(lat, lon);
-            State.UI.CloseDialog();
-        });
-
-        ShowFieldSelectionDialogCommand = new RelayCommand(() =>
-        {
-            // Use settings directory which defaults to ~/Documents/AgValoniaGPS/Fields
-            var fieldsDir = _settingsService.Settings.FieldsDirectory;
-            System.Diagnostics.Debug.WriteLine($"[FieldSelection] Settings.FieldsDirectory = '{fieldsDir}'");
-            if (string.IsNullOrWhiteSpace(fieldsDir))
-            {
-                fieldsDir = Path.Combine(
-                    Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments),
-                    "AgValoniaGPS", "Fields");
-                System.Diagnostics.Debug.WriteLine($"[FieldSelection] Using fallback path: '{fieldsDir}'");
-            }
-            _fieldSelectionDirectory = fieldsDir;
-            System.Diagnostics.Debug.WriteLine($"[FieldSelection] Directory exists: {Directory.Exists(fieldsDir)}");
-
-            // Populate the available fields list
-            PopulateAvailableFields(fieldsDir);
-            System.Diagnostics.Debug.WriteLine($"[FieldSelection] Found {AvailableFields.Count} fields");
-
-            // Show the panel-based dialog
-            State.UI.ShowDialog(DialogType.FieldSelection);
-        });
-
-        CancelFieldSelectionDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            SelectedFieldInfo = null;
-        });
-
-        ConfirmFieldSelectionDialogCommand = new RelayCommand(() =>
-        {
-            if (SelectedFieldInfo == null) return;
-
-            var fieldPath = Path.Combine(_fieldSelectionDirectory, SelectedFieldInfo.Name);
-            FieldsRootDirectory = _fieldSelectionDirectory;
-            CurrentFieldName = SelectedFieldInfo.Name;
-            IsFieldOpen = true;
-
-            // Save as last opened field
-            _settingsService.Settings.LastOpenedField = SelectedFieldInfo.Name;
-            _settingsService.Save();
-
-            // Load field origin from Field.txt (for map centering)
-            try
-            {
-                var fieldInfo = _fieldPlaneFileService.LoadField(fieldPath);
-                if (fieldInfo.Origin != null)
-                {
-                    _fieldOriginLatitude = fieldInfo.Origin.Latitude;
-                    _fieldOriginLongitude = fieldInfo.Origin.Longitude;
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.LogDebug($"[Field] Could not load Field.txt origin: {ex.Message}");
-            }
-
-            // Try to load boundary from field
-            var boundary = _boundaryFileService.LoadBoundary(fieldPath);
-            if (boundary != null)
-            {
-                SetCurrentBoundary(boundary);
-                CenterMapOnBoundary(boundary);
-            }
-
-            // Try to load background image from field
-            LoadBackgroundImage(fieldPath, boundary);
-
-            // Set the active field so headland and other field-specific data loads
-            var field = new Field
-            {
-                Name = SelectedFieldInfo.Name,
-                DirectoryPath = fieldPath,
-                Boundary = boundary
-            };
-            _fieldService.SetActiveField(field);
-
-            // Handle NTRIP profile for this field
-            _ = HandleNtripProfileForFieldAsync(SelectedFieldInfo.Name);
-
-            State.UI.CloseDialog();
-            IsJobMenuPanelVisible = false;
-            StatusMessage = $"Opened field: {SelectedFieldInfo.Name}";
-            SelectedFieldInfo = null;
-        });
-
-        DeleteSelectedFieldCommand = new RelayCommand(() =>
-        {
-            if (SelectedFieldInfo == null) return;
-
-            var fieldPath = Path.Combine(_fieldSelectionDirectory, SelectedFieldInfo.Name);
-            try
-            {
-                if (Directory.Exists(fieldPath))
-                {
-                    Directory.Delete(fieldPath, true);
-                    StatusMessage = $"Deleted field: {SelectedFieldInfo.Name}";
-                    PopulateAvailableFields(_fieldSelectionDirectory);
-                    SelectedFieldInfo = null;
-                }
-            }
-            catch (Exception ex)
-            {
-                StatusMessage = $"Error deleting field: {ex.Message}";
-            }
-        });
-
-        SortFieldsCommand = new RelayCommand(() =>
-        {
-            _fieldsSortedAZ = !_fieldsSortedAZ;
-            var sorted = _fieldsSortedAZ
-                ? AvailableFields.OrderBy(f => f.Name).ToList()
-                : AvailableFields.OrderByDescending(f => f.Name).ToList();
-            AvailableFields.Clear();
-            foreach (var field in sorted)
-            {
-                AvailableFields.Add(field);
-            }
-        });
-
-        ShowNewFieldDialogCommand = new RelayCommand(() =>
-        {
-            // Initialize with current GPS position or defaults
-            NewFieldLatitude = Latitude != 0 ? Latitude : 40.7128;
-            NewFieldLongitude = Longitude != 0 ? Longitude : -74.0060;
-            NewFieldName = string.Empty;
-            State.UI.ShowDialog(DialogType.NewField);
-        });
-
-        CancelNewFieldDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            NewFieldName = string.Empty;
-        });
-
-        ConfirmNewFieldDialogCommand = new RelayCommand(() =>
-        {
-            if (string.IsNullOrWhiteSpace(NewFieldName))
-            {
-                StatusMessage = "Please enter a field name";
-                return;
-            }
-
-            // Get the fields directory
-            var fieldsDir = _settingsService.Settings.FieldsDirectory;
-            if (string.IsNullOrWhiteSpace(fieldsDir))
-            {
-                fieldsDir = Path.Combine(
-                    Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments),
-                    "AgValoniaGPS", "Fields");
-            }
-
-            // Create the field directory
-            var fieldPath = Path.Combine(fieldsDir, NewFieldName);
-            if (Directory.Exists(fieldPath))
-            {
-                StatusMessage = $"Field '{NewFieldName}' already exists";
-                return;
-            }
-
-            try
-            {
-                Directory.CreateDirectory(fieldPath);
-
-                // Save the field origin coordinates
-                var originFile = Path.Combine(fieldPath, "field.origin");
-                File.WriteAllText(originFile, $"{NewFieldLatitude},{NewFieldLongitude}");
-
-                // Create Field.txt in AgOpenGPS format
-                var fieldTxtPath = Path.Combine(fieldPath, "Field.txt");
-                var fieldTxtContent = $"{DateTime.Now:yyyy-MMM-dd hh:mm:ss tt}\n" +
-                                      "$FieldDir\n" +
-                                      $"{NewFieldName}\n" +
-                                      "$Offsets\n" +
-                                      "0,0\n" +
-                                      "Convergence\n" +
-                                      "0\n" +
-                                      "StartFix\n" +
-                                      $"{NewFieldLatitude},{NewFieldLongitude}\n";
-                File.WriteAllText(fieldTxtPath, fieldTxtContent);
-
-                // Set as current field
-                CurrentFieldName = NewFieldName;
-                FieldsRootDirectory = fieldsDir;
-                IsFieldOpen = true;
-
-                // Reset LocalPlane so it will be recreated with new origin
-                _simulatorLocalPlane = null;
-
-                // Save as last opened field
-                _settingsService.Settings.LastOpenedField = NewFieldName;
-                _settingsService.Save();
-
-                State.UI.CloseDialog();
-                IsJobMenuPanelVisible = false;
-                StatusMessage = $"Created field: {NewFieldName}";
-            }
-            catch (Exception ex)
-            {
-                StatusMessage = $"Error creating field: {ex.Message}";
-            }
-        });
-
-        ShowFromExistingFieldDialogCommand = new RelayCommand(() =>
-        {
-            // Populate fields list (reuse same list as field selection)
-            var fieldsDir = _settingsService.Settings.FieldsDirectory;
-            if (string.IsNullOrWhiteSpace(fieldsDir))
-            {
-                fieldsDir = Path.Combine(
-                    Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments),
-                    "AgValoniaGPS", "Fields");
-            }
-            _fieldSelectionDirectory = fieldsDir;
-            PopulateAvailableFields(fieldsDir);
-
-            // Reset copy options
-            CopyFlags = true;
-            CopyMapping = true;
-            CopyHeadland = true;
-            CopyLines = true;
-            FromExistingFieldName = string.Empty;
-            FromExistingSelectedField = null;
-
-            // Pre-select first field if available
-            if (AvailableFields.Count > 0)
-            {
-                FromExistingSelectedField = AvailableFields[0];
-            }
-
-            State.UI.ShowDialog(DialogType.FromExistingField);
-        });
-
-        CancelFromExistingFieldDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            FromExistingSelectedField = null;
-            FromExistingFieldName = string.Empty;
-        });
-
-        ConfirmFromExistingFieldDialogCommand = new RelayCommand(() =>
-        {
-            if (FromExistingSelectedField == null)
-            {
-                StatusMessage = "Please select a field to copy from";
-                return;
-            }
-
-            var newFieldName = FromExistingFieldName.Trim();
-            if (string.IsNullOrWhiteSpace(newFieldName))
-            {
-                StatusMessage = "Please enter a field name";
-                return;
-            }
-
-            // Get the fields directory
-            var fieldsDir = _settingsService.Settings.FieldsDirectory;
-            if (string.IsNullOrWhiteSpace(fieldsDir))
-            {
-                fieldsDir = Path.Combine(
-                    Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments),
-                    "AgValoniaGPS", "Fields");
-            }
-
-            var sourcePath = Path.Combine(fieldsDir, FromExistingSelectedField.Name);
-            var newFieldPath = Path.Combine(fieldsDir, newFieldName);
-
-            // Check if field already exists (unless same name as source)
-            if (Directory.Exists(newFieldPath) && newFieldName != FromExistingSelectedField.Name)
-            {
-                StatusMessage = $"Field '{newFieldName}' already exists";
-                return;
-            }
-
-            try
-            {
-                // Create the new field directory
-                Directory.CreateDirectory(newFieldPath);
-
-                // Copy field.origin if exists
-                var originFile = Path.Combine(sourcePath, "field.origin");
-                if (File.Exists(originFile))
-                {
-                    File.Copy(originFile, Path.Combine(newFieldPath, "field.origin"), true);
-                }
-
-                // Copy boundary
-                var boundaryFile = Path.Combine(sourcePath, "boundary.json");
-                if (File.Exists(boundaryFile))
-                {
-                    File.Copy(boundaryFile, Path.Combine(newFieldPath, "boundary.json"), true);
-                }
-
-                // Copy flags if enabled
-                if (CopyFlags)
-                {
-                    var flagsFile = Path.Combine(sourcePath, "flags.json");
-                    if (File.Exists(flagsFile))
-                    {
-                        File.Copy(flagsFile, Path.Combine(newFieldPath, "flags.json"), true);
-                    }
-                }
-
-                // Copy mapping if enabled
-                if (CopyMapping)
-                {
-                    var mappingFile = Path.Combine(sourcePath, "mapping.json");
-                    if (File.Exists(mappingFile))
-                    {
-                        File.Copy(mappingFile, Path.Combine(newFieldPath, "mapping.json"), true);
-                    }
-                }
-
-                // Copy headland if enabled
-                if (CopyHeadland)
-                {
-                    var headlandFile = Path.Combine(sourcePath, "headland.json");
-                    if (File.Exists(headlandFile))
-                    {
-                        File.Copy(headlandFile, Path.Combine(newFieldPath, "headland.json"), true);
-                    }
-                }
-
-                // Copy lines if enabled
-                if (CopyLines)
-                {
-                    var linesFile = Path.Combine(sourcePath, "lines.json");
-                    if (File.Exists(linesFile))
-                    {
-                        File.Copy(linesFile, Path.Combine(newFieldPath, "lines.json"), true);
-                    }
-                    var abLinesFile = Path.Combine(sourcePath, "ablines.json");
-                    if (File.Exists(abLinesFile))
-                    {
-                        File.Copy(abLinesFile, Path.Combine(newFieldPath, "ablines.json"), true);
-                    }
-                }
-
-                // Set as current field
-                CurrentFieldName = newFieldName;
-                FieldsRootDirectory = fieldsDir;
-                IsFieldOpen = true;
-
-                // Save as last opened field
-                _settingsService.Settings.LastOpenedField = newFieldName;
-                _settingsService.Save();
-
-                State.UI.CloseDialog();
-                IsJobMenuPanelVisible = false;
-                StatusMessage = $"Created field from existing: {newFieldName}";
-            }
-            catch (Exception ex)
-            {
-                StatusMessage = $"Error creating field: {ex.Message}";
-            }
-        });
-
-        AppendVehicleNameCommand = new RelayCommand(() =>
-        {
-            var vehicleName = Vehicle.VehicleTypeDisplayName;
-            if (!string.IsNullOrWhiteSpace(vehicleName))
-            {
-                FromExistingFieldName = (FromExistingFieldName + " " + vehicleName).Trim();
-            }
-        });
-
-        AppendDateCommand = new RelayCommand(() =>
-        {
-            var dateStr = DateTime.Now.ToString("yyyy-MMM-dd");
-            FromExistingFieldName = (FromExistingFieldName + " " + dateStr).Trim();
-        });
-
-        AppendTimeCommand = new RelayCommand(() =>
-        {
-            var timeStr = DateTime.Now.ToString("HH-mm");
-            FromExistingFieldName = (FromExistingFieldName + " " + timeStr).Trim();
-        });
-
-        BackspaceFieldNameCommand = new RelayCommand(() =>
-        {
-            if (FromExistingFieldName.Length > 0)
-            {
-                FromExistingFieldName = FromExistingFieldName.Substring(0, FromExistingFieldName.Length - 1);
-            }
-        });
-
-        ToggleCopyFlagsCommand = new RelayCommand(() => CopyFlags = !CopyFlags);
-        ToggleCopyMappingCommand = new RelayCommand(() => CopyMapping = !CopyMapping);
-        ToggleCopyHeadlandCommand = new RelayCommand(() => CopyHeadland = !CopyHeadland);
-        ToggleCopyLinesCommand = new RelayCommand(() => CopyLines = !CopyLines);
-
-        // KML Import Dialog
-        ShowKmlImportDialogCommand = new RelayCommand(() =>
-        {
-            PopulateAvailableKmlFiles();
-            KmlImportFieldName = string.Empty;
-            KmlBoundaryPointCount = 0;
-            KmlCenterLatitude = 0;
-            KmlCenterLongitude = 0;
-            _kmlBoundaryPoints.Clear();
-            SelectedKmlFile = null;
-
-            // Pre-select first file if available
-            if (AvailableKmlFiles.Count > 0)
-            {
-                SelectedKmlFile = AvailableKmlFiles[0];
-            }
-
-            State.UI.ShowDialog(DialogType.KmlImport);
-        });
-
-        CancelKmlImportDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            SelectedKmlFile = null;
-            KmlImportFieldName = string.Empty;
-        });
-
-        ConfirmKmlImportDialogCommand = new RelayCommand(() =>
-        {
-            if (SelectedKmlFile == null)
-            {
-                StatusMessage = "Please select a KML file";
-                return;
-            }
-
-            var newFieldName = KmlImportFieldName.Trim();
-            if (string.IsNullOrWhiteSpace(newFieldName))
-            {
-                StatusMessage = "Please enter a field name";
-                return;
-            }
-
-            if (_kmlBoundaryPoints.Count < 3)
-            {
-                StatusMessage = "KML file must contain at least 3 boundary points";
-                return;
-            }
-
-            var fieldsDir = _settingsService.Settings.FieldsDirectory;
-            if (string.IsNullOrWhiteSpace(fieldsDir))
-            {
-                fieldsDir = Path.Combine(
-                    Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments),
-                    "AgValoniaGPS", "Fields");
-            }
-
-            var newFieldPath = Path.Combine(fieldsDir, newFieldName);
-            if (Directory.Exists(newFieldPath))
-            {
-                StatusMessage = $"Field '{newFieldName}' already exists";
-                return;
-            }
-
-            try
-            {
-                // Create field directory
-                Directory.CreateDirectory(newFieldPath);
-
-                // Save origin coordinates
-                var originFile = Path.Combine(newFieldPath, "field.origin");
-                File.WriteAllText(originFile, $"{KmlCenterLatitude},{KmlCenterLongitude}");
-
-                // Create and save boundary
-                var origin = new Wgs84(KmlCenterLatitude, KmlCenterLongitude);
-                var sharedProps = new SharedFieldProperties();
-                var localPlane = new LocalPlane(origin, sharedProps);
-
-                var outerPolygon = new BoundaryPolygon();
-                foreach (var (lat, lon) in _kmlBoundaryPoints)
-                {
-                    var wgs84 = new Wgs84(lat, lon);
-                    var geoCoord = localPlane.ConvertWgs84ToGeoCoord(wgs84);
-                    outerPolygon.Points.Add(new BoundaryPoint(geoCoord.Easting, geoCoord.Northing, 0));
-                }
-
-                var boundary = new Boundary { OuterBoundary = outerPolygon };
-                _boundaryFileService.SaveBoundary(boundary, newFieldPath);
-
-                CurrentFieldName = newFieldName;
-                FieldsRootDirectory = fieldsDir;
-                IsFieldOpen = true;
-
-                _settingsService.Settings.LastOpenedField = newFieldName;
-                _settingsService.Save();
-
-                State.UI.CloseDialog();
-                IsJobMenuPanelVisible = false;
-                StatusMessage = $"Imported KML: {newFieldName}";
-            }
-            catch (Exception ex)
-            {
-                StatusMessage = $"Error importing KML: {ex.Message}";
-            }
-        });
-
-        KmlAppendDateCommand = new RelayCommand(() =>
-        {
-            var dateStr = DateTime.Now.ToString("yyyy-MMM-dd");
-            KmlImportFieldName = (KmlImportFieldName + " " + dateStr).Trim();
-        });
-
-        KmlAppendTimeCommand = new RelayCommand(() =>
-        {
-            var timeStr = DateTime.Now.ToString("HH-mm");
-            KmlImportFieldName = (KmlImportFieldName + " " + timeStr).Trim();
-        });
-
-        KmlBackspaceFieldNameCommand = new RelayCommand(() =>
-        {
-            if (KmlImportFieldName.Length > 0)
-            {
-                KmlImportFieldName = KmlImportFieldName.Substring(0, KmlImportFieldName.Length - 1);
-            }
-        });
-
-        // ISO-XML Import Dialog
-        ShowIsoXmlImportDialogCommand = new RelayCommand(() =>
-        {
-            PopulateAvailableIsoXmlFiles();
-            IsoXmlImportFieldName = string.Empty;
-            SelectedIsoXmlFile = null;
-
-            if (AvailableIsoXmlFiles.Count > 0)
-            {
-                SelectedIsoXmlFile = AvailableIsoXmlFiles[0];
-            }
-
-            State.UI.ShowDialog(DialogType.IsoXmlImport);
-        });
-
-        CancelIsoXmlImportDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            SelectedIsoXmlFile = null;
-            IsoXmlImportFieldName = string.Empty;
-        });
-
-        ConfirmIsoXmlImportDialogCommand = new RelayCommand(() =>
-        {
-            if (SelectedIsoXmlFile == null)
-            {
-                StatusMessage = "Please select an ISO-XML folder";
-                return;
-            }
-
-            var newFieldName = IsoXmlImportFieldName.Trim();
-            if (string.IsNullOrWhiteSpace(newFieldName))
-            {
-                StatusMessage = "Please enter a field name";
-                return;
-            }
-
-            var fieldsDir = _settingsService.Settings.FieldsDirectory;
-            if (string.IsNullOrWhiteSpace(fieldsDir))
-            {
-                fieldsDir = Path.Combine(
-                    Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments),
-                    "AgValoniaGPS", "Fields");
-            }
-
-            var newFieldPath = Path.Combine(fieldsDir, newFieldName);
-            if (Directory.Exists(newFieldPath))
-            {
-                StatusMessage = $"Field '{newFieldName}' already exists";
-                return;
-            }
-
-            try
-            {
-                // TODO: Implement ISO-XML parsing when needed
-                // For now, just create the field directory
-                Directory.CreateDirectory(newFieldPath);
-
-                CurrentFieldName = newFieldName;
-                FieldsRootDirectory = fieldsDir;
-                IsFieldOpen = true;
-
-                _settingsService.Settings.LastOpenedField = newFieldName;
-                _settingsService.Save();
-
-                State.UI.CloseDialog();
-                IsJobMenuPanelVisible = false;
-                StatusMessage = $"Imported ISO-XML: {newFieldName}";
-            }
-            catch (Exception ex)
-            {
-                StatusMessage = $"Error importing ISO-XML: {ex.Message}";
-            }
-        });
-
-        IsoXmlAppendDateCommand = new RelayCommand(() =>
-        {
-            var dateStr = DateTime.Now.ToString("yyyy-MMM-dd");
-            IsoXmlImportFieldName = (IsoXmlImportFieldName + " " + dateStr).Trim();
-        });
-
-        IsoXmlAppendTimeCommand = new RelayCommand(() =>
-        {
-            var timeStr = DateTime.Now.ToString("HH-mm");
-            IsoXmlImportFieldName = (IsoXmlImportFieldName + " " + timeStr).Trim();
-        });
-
-        IsoXmlBackspaceFieldNameCommand = new RelayCommand(() =>
-        {
-            if (IsoXmlImportFieldName.Length > 0)
-            {
-                IsoXmlImportFieldName = IsoXmlImportFieldName.Substring(0, IsoXmlImportFieldName.Length - 1);
-            }
-        });
-
-        // Boundary Map Dialog Commands (for satellite map boundary drawing)
-        ShowBoundaryMapDialogCommand = new RelayCommand(() =>
-        {
-            // Set center: prefer field origin, then GPS position, then 0,0
-            if (_fieldOriginLatitude != 0 || _fieldOriginLongitude != 0)
-            {
-                // Use field origin from Field.txt
-                BoundaryMapCenterLatitude = _fieldOriginLatitude;
-                BoundaryMapCenterLongitude = _fieldOriginLongitude;
-            }
-            else if (Latitude != 0 || Longitude != 0)
-            {
-                // Fall back to current GPS position
-                BoundaryMapCenterLatitude = Latitude;
-                BoundaryMapCenterLongitude = Longitude;
-            }
-            // else: leave as 0,0 (will show default location)
-
-            BoundaryMapPointCount = 0;
-            BoundaryMapCanSave = false;
-            BoundaryMapCoordinateText = string.Empty;
-            BoundaryMapResultPoints.Clear();
-            State.UI.ShowDialog(DialogType.BoundaryMap);
-        });
-
-        CancelBoundaryMapDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            BoundaryMapResultPoints.Clear();
-        });
-
-        ConfirmBoundaryMapDialogCommand = new RelayCommand(() =>
-        {
-            _logger.LogDebug($"[BoundaryMap] ConfirmBoundaryMapDialogCommand called");
-            _logger.LogDebug($"[BoundaryMap] Points: {BoundaryMapResultPoints.Count}, IsFieldOpen: {IsFieldOpen}, CurrentFieldName: {CurrentFieldName}");
-
-            if (BoundaryMapResultPoints.Count >= 3 && IsFieldOpen && !string.IsNullOrEmpty(CurrentFieldName))
-            {
-                try
-                {
-                    var fieldPath = Path.Combine(_settingsService.Settings.FieldsDirectory, CurrentFieldName);
-                    _logger.LogDebug($"[BoundaryMap] Field path: {fieldPath}");
-                    _logger.LogDebug($"[BoundaryMap] Directory exists: {Directory.Exists(fieldPath)}");
-
-                    // Load existing boundary or create new one
-                    var boundary = _boundaryFileService.LoadBoundary(fieldPath) ?? new Boundary();
-
-                    // Calculate origin for LocalPlane
-                    // IMPORTANT: If we have a background image, use its center as the origin
-                    // This ensures the boundary aligns with landmarks in the image
-                    // The user drew the boundary on specific landmarks in the viewport,
-                    // so we need to use the same reference point for both
-                    double centerLat, centerLon;
-                    if (!string.IsNullOrEmpty(BoundaryMapResultBackgroundPath))
-                    {
-                        // Use image (viewport) center as origin - this is where the user was looking when drawing
-                        centerLat = (BoundaryMapResultNwLat + BoundaryMapResultSeLat) / 2;
-                        centerLon = (BoundaryMapResultNwLon + BoundaryMapResultSeLon) / 2;
-                        _logger.LogDebug($"[BoundaryMap] Using image center as origin: ({centerLat:F8}, {centerLon:F8})");
-                    }
-                    else
-                    {
-                        // No background image - use boundary center as origin
-                        centerLat = BoundaryMapResultPoints.Average(p => p.Latitude);
-                        centerLon = BoundaryMapResultPoints.Average(p => p.Longitude);
-                        _logger.LogDebug($"[BoundaryMap] Using boundary center as origin: ({centerLat:F8}, {centerLon:F8})");
-                    }
-
-                    // Convert WGS84 boundary points to local coordinates
-                    var origin = new Wgs84(centerLat, centerLon);
-                    var sharedProps = new SharedFieldProperties();
-                    var localPlane = new LocalPlane(origin, sharedProps);
-
-                    var outerPolygon = new BoundaryPolygon();
-
-                    _logger.LogDebug($"[BoundaryMap] Converting boundary points with origin ({centerLat:F8}, {centerLon:F8})");
-                    foreach (var (lat, lon) in BoundaryMapResultPoints)
-                    {
-                        var wgs84 = new Wgs84(lat, lon);
-                        var geoCoord = localPlane.ConvertWgs84ToGeoCoord(wgs84);
-                        outerPolygon.Points.Add(new BoundaryPoint(geoCoord.Easting, geoCoord.Northing, 0));
-                        _logger.LogDebug($"[BoundaryMap]   WGS84 ({lat:F8}, {lon:F8}) -> Local E={geoCoord.Easting:F1}, N={geoCoord.Northing:F1}");
-                    }
-
-                    boundary.OuterBoundary = outerPolygon;
-
-                    // Save boundary
-                    _boundaryFileService.SaveBoundary(boundary, fieldPath);
-
-                    // Update Field.txt with the origin used for this boundary
-                    // This ensures background images load with the same coordinate system
-                    _fieldOriginLatitude = centerLat;
-                    _fieldOriginLongitude = centerLon;
-                    try
-                    {
-                        var fieldInfo = _fieldPlaneFileService.LoadField(fieldPath);
-                        fieldInfo.Origin = new Position { Latitude = centerLat, Longitude = centerLon };
-                        _fieldPlaneFileService.SaveField(fieldInfo, fieldPath);
-                        _logger.LogDebug($"[BoundaryMap] Updated Field.txt origin to ({centerLat:F8}, {centerLon:F8})");
-                    }
-                    catch (Exception ex)
-                    {
-                        _logger.LogDebug($"[BoundaryMap] Could not update Field.txt: {ex.Message}");
-                    }
-
-                    // Update map
-                    SetCurrentBoundary(boundary);
-
-                    // Center camera on the boundary and set appropriate zoom
-                    if (outerPolygon.Points.Count > 0)
-                    {
-                        // Calculate boundary center and extent
-                        double minE = double.MaxValue, maxE = double.MinValue;
-                        double minN = double.MaxValue, maxN = double.MinValue;
-                        foreach (var pt in outerPolygon.Points)
-                        {
-                            minE = Math.Min(minE, pt.Easting);
-                            maxE = Math.Max(maxE, pt.Easting);
-                            minN = Math.Min(minN, pt.Northing);
-                            maxN = Math.Max(maxN, pt.Northing);
-                        }
-                        double centerE = (minE + maxE) / 2.0;
-                        double centerN = (minN + maxN) / 2.0;
-                        double extentE = maxE - minE;
-                        double extentN = maxN - minN;
-                        double maxExtent = Math.Max(extentE, extentN);
-
-                        // Pan to center
-                        _mapService.PanTo(centerE, centerN);
-
-                        // Calculate zoom to fit boundary (viewHeight = 200/zoom, so zoom = 200/viewHeight)
-                        // Add 20% padding
-                        double desiredView = maxExtent * 1.2;
-                        if (desiredView > 0)
-                        {
-                            double newZoom = 200.0 / desiredView;
-                            newZoom = Math.Clamp(newZoom, 0.1, 10.0);
-                            _mapService.SetCamera(centerE, centerN, newZoom, 0);
-                        }
-
-                        _logger.LogDebug($"[BoundaryMap] Saved boundary with {outerPolygon.Points.Count} points");
-                        _logger.LogDebug($"[BoundaryMap] Center: ({centerE:F1}, {centerN:F1}), Extent: {maxExtent:F1}m");
-                    }
-
-                    // Handle background image if captured
-                    if (!string.IsNullOrEmpty(BoundaryMapResultBackgroundPath))
-                    {
-                        SaveBackgroundImage(BoundaryMapResultBackgroundPath, fieldPath,
-                            BoundaryMapResultNwLat, BoundaryMapResultNwLon,
-                            BoundaryMapResultSeLat, BoundaryMapResultSeLon);
-                    }
-
-                    // Refresh the boundary list
-                    RefreshBoundaryList();
-
-                    StatusMessage = $"Boundary created with {BoundaryMapResultPoints.Count} points";
-                }
-                catch (Exception ex)
-                {
-                    StatusMessage = $"Error creating boundary: {ex.Message}";
-                }
-            }
-
-            State.UI.CloseDialog();
-            IsBoundaryPanelVisible = false;
-            BoundaryMapResultPoints.Clear();
-        });
-
-        ShowAgShareDownloadDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.ShowDialog(DialogType.AgShareDownload);
-        });
-
-        CancelAgShareDownloadDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-        });
-
-        ShowAgShareUploadDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.ShowDialog(DialogType.AgShareUpload);
-        });
-
-        CancelAgShareUploadDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-        });
-
-        ShowAgShareSettingsDialogCommand = new RelayCommand(() =>
-        {
-            // Load current settings from storage
-            AgShareSettingsServerUrl = _settingsService.Settings.AgShareServer;
-            AgShareSettingsApiKey = _settingsService.Settings.AgShareApiKey;
-            AgShareSettingsEnabled = _settingsService.Settings.AgShareEnabled;
-            State.UI.ShowDialog(DialogType.AgShareSettings);
-        });
-
-        CancelAgShareSettingsDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-        });
-
-        ConfirmAgShareSettingsDialogCommand = new RelayCommand(() =>
-        {
-            // Save settings to storage
-            _settingsService.Settings.AgShareServer = AgShareSettingsServerUrl;
-            _settingsService.Settings.AgShareApiKey = AgShareSettingsApiKey;
-            _settingsService.Settings.AgShareEnabled = AgShareSettingsEnabled;
-            _settingsService.Save();
-
-            State.UI.CloseDialog();
-            StatusMessage = "AgShare settings saved";
-        });
-
-        ShowBoundaryDialogCommand = new RelayCommand(() =>
-        {
-            // Toggle boundary panel visibility - this shows the panel where user can
-            // choose how to create the boundary (KML import, drive around, etc.)
-            IsBoundaryPanelVisible = !IsBoundaryPanelVisible;
-        });
-
-        // Headland Commands
-        ShowHeadlandBuilderCommand = new RelayCommand(() =>
-        {
-            if (!IsFieldOpen)
-            {
-                StatusMessage = "Open a field first";
-                return;
-            }
-            State.UI.ShowDialog(DialogType.HeadlandBuilder);
-            // Trigger initial preview
-            UpdateHeadlandPreview();
-        });
-
-        ToggleHeadlandCommand = new RelayCommand(() =>
-        {
-            if (!HasHeadland)
-            {
-                StatusMessage = "No headland defined";
-                return;
-            }
-            IsHeadlandOn = !IsHeadlandOn;
-        });
-
-        ToggleSectionInHeadlandCommand = new RelayCommand(() =>
-        {
-            IsSectionControlInHeadland = !IsSectionControlInHeadland;
-            StatusMessage = IsSectionControlInHeadland ? "Section control in headland: ON" : "Section control in headland: OFF";
-        });
-
-        ResetToolHeadingCommand = new RelayCommand(() =>
-        {
-            // Reset tool heading to be directly behind tractor
-            StatusMessage = "Tool heading reset";
-        });
-
-        BuildHeadlandCommand = new RelayCommand(() =>
-        {
-            BuildHeadlandFromBoundary();
-        });
-
-        ClearHeadlandCommand = new RelayCommand(() =>
-        {
-            CurrentHeadlandLine = null;
-            HeadlandPreviewLine = null;
-            HasHeadland = false;
-            IsHeadlandOn = false;
-            StatusMessage = "Headland cleared";
-        });
-
-        CloseHeadlandBuilderCommand = new RelayCommand(() =>
-        {
-            HeadlandPreviewLine = null;
-            State.UI.CloseDialog();
-        });
-
-        SetHeadlandToToolWidthCommand = new RelayCommand(() =>
-        {
-            // Set headland distance to implement width (use actual width from sections)
-            double actualWidth = ConfigStore.ActualToolWidth;
-            HeadlandDistance = actualWidth > 0 ? actualWidth * 2 : 12.0;
-            UpdateHeadlandPreview();
-        });
-
-        PreviewHeadlandCommand = new RelayCommand(() =>
-        {
-            UpdateHeadlandPreview();
-        });
-
-        IncrementHeadlandDistanceCommand = new RelayCommand(() =>
-        {
-            HeadlandDistance = Math.Min(HeadlandDistance + 0.5, 100.0);
-            UpdateHeadlandPreview();
-        });
-
-        DecrementHeadlandDistanceCommand = new RelayCommand(() =>
-        {
-            HeadlandDistance = Math.Max(HeadlandDistance - 0.5, 0.5);
-            UpdateHeadlandPreview();
-        });
-
-        IncrementHeadlandPassesCommand = new RelayCommand(() =>
-        {
-            HeadlandPasses = Math.Min(HeadlandPasses + 1, 10);
-            UpdateHeadlandPreview();
-        });
-
-        DecrementHeadlandPassesCommand = new RelayCommand(() =>
-        {
-            HeadlandPasses = Math.Max(HeadlandPasses - 1, 1);
-            UpdateHeadlandPreview();
-        });
-
-        // Headland Dialog (FormHeadLine) commands
-        ShowHeadlandDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.ShowDialog(DialogType.Headland);
-            UpdateHeadlandPreview();
-        });
-
-        CloseHeadlandDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            HeadlandPreviewLine = null;
-        });
-
-        ExtendHeadlandACommand = new RelayCommand(() =>
-        {
-            // TODO: Extend headland at point A
-            StatusMessage = "Extend A - not yet implemented";
-        });
-
-        ExtendHeadlandBCommand = new RelayCommand(() =>
-        {
-            // TODO: Extend headland at point B
-            StatusMessage = "Extend B - not yet implemented";
-        });
-
-        ShrinkHeadlandACommand = new RelayCommand(() =>
-        {
-            // TODO: Shrink headland at point A
-            StatusMessage = "Shrink A - not yet implemented";
-        });
-
-        ShrinkHeadlandBCommand = new RelayCommand(() =>
-        {
-            // TODO: Shrink headland at point B
-            StatusMessage = "Shrink B - not yet implemented";
-        });
-
-        ResetHeadlandCommand = new RelayCommand(() =>
-        {
-            ClearHeadlandCommand?.Execute(null);
-            StatusMessage = "Headland reset";
-        });
-
-        ClipHeadlandLineCommand = new RelayCommand(() =>
-        {
-            if (!HeadlandPointsSelected)
-            {
-                StatusMessage = "Select 2 points on the boundary first";
-                return;
-            }
-
-            // Check if we have a headland to clip (either built headland or preview)
-            var headlandToClip = CurrentHeadlandLine ?? ConvertPreviewToVec3(HeadlandPreviewLine);
-            if (headlandToClip == null || headlandToClip.Count < 3)
-            {
-                StatusMessage = "No headland to clip - use Build first";
-                return;
-            }
-
-            // Clip the headland using the clip line (between the two selected points)
-            ClipHeadlandAtLine(headlandToClip);
-        });
-
-        UndoHeadlandCommand = new RelayCommand(() =>
-        {
-            // TODO: Undo headland changes
-            StatusMessage = "Undo - not yet implemented";
-        });
-
-        TurnOffHeadlandCommand = new RelayCommand(() =>
-        {
-            IsHeadlandOn = false;
-            HasHeadland = false;
-            CurrentHeadlandLine = null;
-            HeadlandPreviewLine = null;
-            StatusMessage = "Headland turned off";
-        });
-
-        // AB Line Guidance Commands - Bottom Bar
-        SnapLeftCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Snap to Left Track - not yet implemented";
-        });
-
-        SnapRightCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Snap to Right Track - not yet implemented";
-        });
-
-        StopGuidanceCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Guidance Stopped";
-        });
-
-        UTurnCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "U-Turn - not yet implemented";
-        });
-
-        // AB Line Guidance Commands - Flyout Menu
-        ShowTracksDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.ShowDialog(DialogType.Tracks);
-        });
-
-        CloseTracksDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-        });
-
-        // Track management commands
-        DeleteSelectedTrackCommand = new RelayCommand(() =>
-        {
-            if (SelectedTrack != null)
-            {
-                SavedTracks.Remove(SelectedTrack);
-                SelectedTrack = null;
-                SaveTracksToFile(); // Persist deletion to disk
-                StatusMessage = "Track deleted";
-            }
-        });
-
-        SwapABPointsCommand = new RelayCommand(() =>
-        {
-            if (SelectedTrack != null && SelectedTrack.Points.Count >= 2)
-            {
-                // Reverse the points list to swap A and B
-                SelectedTrack.Points.Reverse();
-                StatusMessage = $"Swapped A/B points for {SelectedTrack.Name}";
-            }
-        });
-
-        SelectTrackAsActiveCommand = new RelayCommand(() =>
-        {
-            // Toggle: if track is already active, deactivate it; otherwise activate
-            if (SelectedTrack != null)
-            {
-                if (SelectedTrack.IsActive)
-                {
-                    // Deactivate
-                    SelectedTrack = null;
-                    StatusMessage = "Track deactivated";
-                }
-                else
-                {
-                    // Activate (SelectedTrack setter handles IsActive sync)
-                    StatusMessage = $"Activated track: {SelectedTrack.Name}";
-                }
-                State.UI.CloseDialog();
-            }
-        });
-
-        // NTRIP Profiles commands
-        ShowNtripProfilesDialogCommand = new RelayCommand(() =>
-        {
-            RefreshNtripProfiles();
-            State.UI.ShowDialog(DialogType.NtripProfiles);
-        });
-
-        CloseNtripProfilesDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            SelectedNtripProfile = null;
-        });
-
-        AddNtripProfileCommand = new RelayCommand(() =>
-        {
-            EditingNtripProfile = _ntripProfileService.CreateNewProfile("New Profile");
-            PopulateAvailableFieldsForProfile(EditingNtripProfile);
-            State.UI.ShowDialog(DialogType.NtripProfileEditor);
-        });
-
-        EditNtripProfileCommand = new RelayCommand(() =>
-        {
-            if (SelectedNtripProfile != null)
-            {
-                // Clone the profile for editing
-                EditingNtripProfile = new NtripProfile
-                {
-                    Id = SelectedNtripProfile.Id,
-                    Name = SelectedNtripProfile.Name,
-                    CasterHost = SelectedNtripProfile.CasterHost,
-                    CasterPort = SelectedNtripProfile.CasterPort,
-                    MountPoint = SelectedNtripProfile.MountPoint,
-                    Username = SelectedNtripProfile.Username,
-                    Password = SelectedNtripProfile.Password,
-                    AutoConnectOnFieldLoad = SelectedNtripProfile.AutoConnectOnFieldLoad,
-                    IsDefault = SelectedNtripProfile.IsDefault,
-                    AssociatedFields = new List<string>(SelectedNtripProfile.AssociatedFields),
-                    FilePath = SelectedNtripProfile.FilePath
-                };
-                PopulateAvailableFieldsForProfile(EditingNtripProfile);
-                State.UI.ShowDialog(DialogType.NtripProfileEditor);
-            }
-        });
-
-        DeleteNtripProfileCommand = new AsyncRelayCommand(async () =>
-        {
-            if (SelectedNtripProfile != null)
-            {
-                var confirmed = await _dialogService.ShowConfirmationAsync(
-                    "Delete NTRIP Profile",
-                    $"Are you sure you want to delete the profile '{SelectedNtripProfile.Name}'?");
-                if (!confirmed) return;
-
-                await _ntripProfileService.DeleteProfileAsync(SelectedNtripProfile.Id);
-                RefreshNtripProfiles();
-                SelectedNtripProfile = null;
-                StatusMessage = "NTRIP profile deleted";
-            }
-        });
-
-        SetDefaultNtripProfileCommand = new AsyncRelayCommand(async () =>
-        {
-            if (SelectedNtripProfile != null)
-            {
-                await _ntripProfileService.SetDefaultProfileAsync(SelectedNtripProfile.Id);
-                RefreshNtripProfiles();
-                StatusMessage = $"Set '{SelectedNtripProfile.Name}' as default NTRIP profile";
-            }
-        });
-
-        SaveNtripProfileCommand = new AsyncRelayCommand(async () =>
-        {
-            if (EditingNtripProfile != null)
-            {
-                // Update associated fields from the selection
-                EditingNtripProfile.AssociatedFields = AvailableFieldsForProfile
-                    .Where(f => f.IsSelected)
-                    .Select(f => f.FieldName)
-                    .ToList();
-
-                await _ntripProfileService.SaveProfileAsync(EditingNtripProfile);
-                RefreshNtripProfiles();
-                EditingNtripProfile = null;
-                AvailableFieldsForProfile.Clear();
-                State.UI.ShowDialog(DialogType.NtripProfiles);
-                StatusMessage = "NTRIP profile saved";
-            }
-        });
-
-        CancelNtripProfileEditCommand = new RelayCommand(() =>
-        {
-            EditingNtripProfile = null;
-            AvailableFieldsForProfile.Clear();
-            NtripTestStatus = string.Empty;
-            State.UI.ShowDialog(DialogType.NtripProfiles);
-        });
-
-        TestNtripConnectionCommand = new AsyncRelayCommand(async () =>
-        {
-            if (EditingNtripProfile == null) return;
-            if (string.IsNullOrWhiteSpace(EditingNtripProfile.CasterHost))
-            {
-                NtripTestStatus = "Error: Caster host is required";
-                return;
-            }
-            if (string.IsNullOrWhiteSpace(EditingNtripProfile.MountPoint))
-            {
-                NtripTestStatus = "Error: Mount point is required";
-                return;
-            }
-
-            IsTestingNtripConnection = true;
-            NtripTestStatus = "Testing connection...";
-
-            try
-            {
-                using var tcpClient = new System.Net.Sockets.TcpClient();
-                var connectTask = tcpClient.ConnectAsync(
-                    EditingNtripProfile.CasterHost,
-                    EditingNtripProfile.CasterPort);
-
-                if (await Task.WhenAny(connectTask, Task.Delay(5000)) == connectTask)
-                {
-                    if (tcpClient.Connected)
-                    {
-                        // Try to get the mount point table to verify caster is responding
-                        using var stream = tcpClient.GetStream();
-                        var request = $"GET /{EditingNtripProfile.MountPoint} HTTP/1.1\r\n" +
-                                    $"Host: {EditingNtripProfile.CasterHost}\r\n" +
-                                    $"Ntrip-Version: Ntrip/2.0\r\n" +
-                                    $"User-Agent: NTRIP AgValoniaGPS/Test\r\n";
-
-                        if (!string.IsNullOrEmpty(EditingNtripProfile.Username))
-                        {
-                            var credentials = Convert.ToBase64String(
-                                System.Text.Encoding.ASCII.GetBytes(
-                                    $"{EditingNtripProfile.Username}:{EditingNtripProfile.Password}"));
-                            request += $"Authorization: Basic {credentials}\r\n";
-                        }
-                        request += "\r\n";
-
-                        var requestBytes = System.Text.Encoding.ASCII.GetBytes(request);
-                        await stream.WriteAsync(requestBytes, 0, requestBytes.Length);
-
-                        // Read response with timeout
-                        var buffer = new byte[1024];
-                        stream.ReadTimeout = 3000;
-                        var bytesRead = await stream.ReadAsync(buffer, 0, buffer.Length);
-                        var response = System.Text.Encoding.ASCII.GetString(buffer, 0, bytesRead);
-
-                        if (response.Contains("200 OK") || response.Contains("ICY 200"))
-                        {
-                            NtripTestStatus = "Success: Connected to caster and mount point";
-                        }
-                        else if (response.Contains("401"))
-                        {
-                            NtripTestStatus = "Error: Authentication failed (check username/password)";
-                        }
-                        else if (response.Contains("404"))
-                        {
-                            NtripTestStatus = "Error: Mount point not found";
-                        }
-                        else
-                        {
-                            NtripTestStatus = "Connected to caster (mount point status unknown)";
-                        }
-                    }
-                    else
-                    {
-                        NtripTestStatus = "Error: Could not connect to caster";
-                    }
-                }
-                else
-                {
-                    NtripTestStatus = "Error: Connection timed out";
-                }
-            }
-            catch (System.Net.Sockets.SocketException ex)
-            {
-                NtripTestStatus = $"Error: {ex.Message}";
-            }
-            catch (Exception ex)
-            {
-                NtripTestStatus = $"Error: {ex.Message}";
-            }
-            finally
-            {
-                IsTestingNtripConnection = false;
-            }
-        });
-
-        ShowQuickABSelectorCommand = new RelayCommand(() =>
-        {
-            State.UI.ShowDialog(DialogType.QuickABSelector);
-        });
-
-        CloseQuickABSelectorCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-        });
-
-        ShowDrawABDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.ShowDialog(DialogType.DrawAB);
-        });
-
-        CloseDrawABDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-        });
-
-        StartNewABLineCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Starting new AB Line - not yet implemented";
-        });
-
-        StartNewABCurveCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Starting new AB Curve - not yet implemented";
-        });
-
-        // Quick AB Mode Commands
-        StartAPlusLineCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            StatusMessage = "A+ Line mode: Line created from current position and heading";
-            // TODO: Create AB line from current position using current heading
-        });
-
-        StartDriveABCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            CurrentABCreationMode = ABCreationMode.DriveAB;
-            CurrentABPointStep = ABPointStep.SettingPointA;
-            PendingPointA = null;
-            StatusMessage = ABCreationInstructions;
-        });
-
-        StartCurveRecordingCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            StatusMessage = "Curve mode: Start driving to record curve path";
-            // TODO: Start curve recording mode
-        });
-
-        StartDrawABModeCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            CurrentABCreationMode = ABCreationMode.DrawAB;
-            CurrentABPointStep = ABPointStep.SettingPointA;
-            PendingPointA = null;
-            StatusMessage = ABCreationInstructions;
-        });
-
-        // SetABPointCommand is called when user taps during AB creation mode
-        // For DriveAB mode: uses current GPS position
-        // For DrawAB mode: uses the tapped map coordinates (passed as parameter)
-        SetABPointCommand = new RelayCommand<object?>(param =>
-        {
-            _logger.LogDebug($"[SetABPointCommand] Called with param={param?.GetType().Name ?? "null"}, Mode={CurrentABCreationMode}, Step={CurrentABPointStep}");
-
-            if (CurrentABCreationMode == ABCreationMode.None)
-            {
-                _logger.LogDebug("[SetABPointCommand] Mode is None, returning");
-                return;
-            }
-
-            Position pointToSet;
-
-            if (CurrentABCreationMode == ABCreationMode.DriveAB)
-            {
-                // Use current GPS position
-                pointToSet = new Position
-                {
-                    Latitude = Latitude,
-                    Longitude = Longitude,
-                    Easting = Easting,
-                    Northing = Northing,
-                    Heading = Heading
-                };
-                _logger.LogDebug($"[SetABPointCommand] DriveAB - GPS position: E={Easting:F2}, N={Northing:F2}");
-            }
-            else if (CurrentABCreationMode == ABCreationMode.DrawAB && param is Position mapPos)
-            {
-                // Use the tapped map position
-                pointToSet = mapPos;
-                _logger.LogDebug($"[SetABPointCommand] DrawAB - Map position: E={mapPos.Easting:F2}, N={mapPos.Northing:F2}");
-            }
-            else
-            {
-                _logger.LogDebug($"[SetABPointCommand] Invalid state - returning");
-                return; // Invalid state
-            }
-
-            if (CurrentABPointStep == ABPointStep.SettingPointA)
-            {
-                // Store Point A and move to Point B
-                PendingPointA = pointToSet;
-                CurrentABPointStep = ABPointStep.SettingPointB;
-                StatusMessage = ABCreationInstructions;
-                _logger.LogDebug($"[SetABPointCommand] Set Point A: E={pointToSet.Easting:F2}, N={pointToSet.Northing:F2}");
-            }
-            else if (CurrentABPointStep == ABPointStep.SettingPointB)
-            {
-                // Create the AB line with Point A and Point B
-                if (PendingPointA != null)
-                {
-                    var heading = CalculateHeading(PendingPointA, pointToSet);
-                    var headingRadians = heading * Math.PI / 180.0;
-                    var newTrack = Track.FromABLine(
-                        $"AB_{heading:F1}° {DateTime.Now:HH:mm:ss}",
-                        new Vec3(PendingPointA.Easting, PendingPointA.Northing, headingRadians),
-                        new Vec3(pointToSet.Easting, pointToSet.Northing, headingRadians));
-                    newTrack.IsActive = true;
-
-                    SavedTracks.Add(newTrack);
-                    SaveTracksToFile(); // Persist to disk
-                    HasActiveTrack = true;
-                    IsAutoSteerAvailable = true;
-                    StatusMessage = $"Created AB line: {newTrack.Name} ({heading:F1}°)";
-                    _logger.LogDebug($"[SetABPointCommand] Created AB Line: {newTrack.Name}, A=({PendingPointA.Easting:F2},{PendingPointA.Northing:F2}), B=({pointToSet.Easting:F2},{pointToSet.Northing:F2}), Heading={heading:F1}°");
-
-                    // Reset state
-                    CurrentABCreationMode = ABCreationMode.None;
-                    CurrentABPointStep = ABPointStep.None;
-                    PendingPointA = null;
-                }
-            }
-        });
-
-        CancelABCreationCommand = new RelayCommand(() =>
-        {
-            CurrentABCreationMode = ABCreationMode.None;
-            CurrentABPointStep = ABPointStep.None;
-            PendingPointA = null;
-            StatusMessage = "AB line creation cancelled";
-        });
-
-        CycleABLinesCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Cycle AB Lines - not yet implemented";
-        });
-
-        SmoothABLineCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Smooth AB Line - not yet implemented";
-        });
-
-        NudgeLeftCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Nudge Left - not yet implemented";
-        });
-
-        NudgeRightCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Nudge Right - not yet implemented";
-        });
-
-        FineNudgeLeftCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Fine Nudge Left - not yet implemented";
-        });
-
-        FineNudgeRightCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Fine Nudge Right - not yet implemented";
-        });
-
-        // Bottom Strip Commands (matching AgOpenGPS panelBottom)
-        ChangeMappingColorCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Section Mapping Color - not yet implemented";
-        });
-
-        SnapToPivotCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Snap to Pivot - not yet implemented";
-        });
-
-        ToggleYouSkipCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "YouSkip Toggle - not yet implemented";
-        });
-
-        ToggleUTurnSkipRowsCommand = new RelayCommand(() =>
-        {
-            IsUTurnSkipRowsEnabled = !IsUTurnSkipRowsEnabled;
-            StatusMessage = IsUTurnSkipRowsEnabled
-                ? $"U-Turn skip rows: ON ({UTurnSkipRows} rows)"
-                : "U-Turn skip rows: OFF";
-        });
-
-        CycleUTurnSkipRowsCommand = new RelayCommand(() =>
-        {
-            // Cycle through 0-9, wrap back to 0 after 9
-            UTurnSkipRows = (UTurnSkipRows + 1) % 10;
-            StatusMessage = $"Skip rows: {UTurnSkipRows}";
-        });
-
-        // Flags Commands
-        PlaceRedFlagCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Place Red Flag - not yet implemented";
-        });
-
-        PlaceGreenFlagCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Place Green Flag - not yet implemented";
-        });
-
-        PlaceYellowFlagCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Place Yellow Flag - not yet implemented";
-        });
-
-        DeleteAllFlagsCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Delete All Flags - not yet implemented";
-        });
-
-        // Right Navigation Panel Commands
-        ToggleContourModeCommand = new RelayCommand(() =>
-        {
-            IsContourModeOn = !IsContourModeOn;
-            StatusMessage = IsContourModeOn ? "Contour mode ON" : "Contour mode OFF";
-        });
-
-        DeleteContoursCommand = new RelayCommand(() =>
-        {
-            _coverageMapService.ClearAll();
-            StatusMessage = "Coverage/contours cleared";
-        });
-
-        DeleteAppliedAreaCommand = new RelayCommand(async () =>
-        {
-            // Show confirmation dialog
-            var confirmed = await _dialogService.ShowConfirmationAsync(
-                "Delete Applied Area",
-                "Are you sure you want to delete all applied area coverage? This cannot be undone.");
-
-            if (!confirmed)
-                return;
-
-            // Clear in-memory coverage
-            _coverageMapService.ClearAll();
-
-            // Delete Sections.txt file from current field
-            if (State.Field.ActiveField != null)
-            {
-                var sectionsFile = System.IO.Path.Combine(State.Field.ActiveField.DirectoryPath, "Sections.txt");
-                if (System.IO.File.Exists(sectionsFile))
-                {
-                    try
-                    {
-                        System.IO.File.Delete(sectionsFile);
-                        _logger.LogDebug($"[Coverage] Deleted {sectionsFile}");
-                    }
-                    catch (Exception ex)
-                    {
-                        _logger.LogDebug($"[Coverage] Error deleting Sections.txt: {ex.Message}");
-                    }
-                }
-            }
-
-            // Refresh statistics display
-            RefreshCoverageStatistics();
-            StatusMessage = "Applied area deleted";
-        });
-
-        // Manual All button: Toggle all sections between On (green) and Off (red)
-        ToggleManualModeCommand = new RelayCommand(() =>
-        {
-            IsManualSectionMode = !IsManualSectionMode;
-
-            // Turn off Auto button if turning on Manual
-            if (IsManualSectionMode)
-                IsSectionMasterOn = false;
-
-            // Set all sections
-            var newState = IsManualSectionMode ? SectionButtonState.On : SectionButtonState.Off;
-            for (int i = 0; i < _sectionControlService.NumSections; i++)
-            {
-                _sectionControlService.SetSectionState(i, newState);
-            }
-
-            StatusMessage = IsManualSectionMode ? "All sections ON" : "All sections OFF";
-        });
-
-        // Auto All button: Toggle all sections between Auto (yellow) and Off (red)
-        ToggleSectionMasterCommand = new RelayCommand(() =>
-        {
-            IsSectionMasterOn = !IsSectionMasterOn;
-
-            // Turn off Manual button if turning on Auto
-            if (IsSectionMasterOn)
-                IsManualSectionMode = false;
-
-            // Set all sections
-            var newState = IsSectionMasterOn ? SectionButtonState.Auto : SectionButtonState.Off;
-            for (int i = 0; i < _sectionControlService.NumSections; i++)
-            {
-                _sectionControlService.SetSectionState(i, newState);
-            }
-
-            StatusMessage = IsSectionMasterOn ? "All sections AUTO" : "All sections OFF";
-        });
-
-        // Parameterized command for toggling individual sections (0-based index)
-        // Parameter comes as string from XAML CommandParameter, so we parse it
-        ToggleSectionCommand = new RelayCommand<object>(param =>
-        {
-            if (param == null) return;
-
-            int sectionIndex;
-            if (param is int intVal)
-                sectionIndex = intVal;
-            else if (param is string strVal && int.TryParse(strVal, out var parsed))
-                sectionIndex = parsed;
-            else
-                return;
-
-            if (sectionIndex < 0 || sectionIndex >= _sectionControlService.NumSections)
-                return;
-
-            var currentState = _sectionControlService.SectionStates[sectionIndex].ButtonState;
-            // Cycle through: Off -> Auto -> On -> Off
-            var newState = currentState switch
-            {
-                SectionButtonState.Off => SectionButtonState.Auto,
-                SectionButtonState.Auto => SectionButtonState.On,
-                SectionButtonState.On => SectionButtonState.Off,
-                _ => SectionButtonState.Off
-            };
-
-            _sectionControlService.SetSectionState(sectionIndex, newState);
-            StatusMessage = $"Section {sectionIndex + 1}: {newState}";
-        });
-
-        ToggleYouTurnCommand = new RelayCommand(() =>
-        {
-            IsYouTurnEnabled = !IsYouTurnEnabled;
-            StatusMessage = IsYouTurnEnabled ? "YouTurn enabled" : "YouTurn disabled";
-        });
-
-        ToggleAutoSteerCommand = new RelayCommand(() =>
-        {
-            if (!IsAutoSteerAvailable)
-            {
-                StatusMessage = "AutoSteer not available - no active track";
-                return;
-            }
-            IsAutoSteerEngaged = !IsAutoSteerEngaged;
-            StatusMessage = IsAutoSteerEngaged ? "AutoSteer ENGAGED" : "AutoSteer disengaged";
-        });
-
-        // Field Commands
-        CloseFieldCommand = new AsyncRelayCommand(async () =>
-        {
-            CurrentFieldName = string.Empty;
-            IsFieldOpen = false;
-            SetCurrentBoundary(null);
-
-            // Clear headland
-            LoadHeadlandFromField(null);
-
-            // Clear tracks
-            State.Field.Tracks.Clear();
-            SavedTracks.Clear();
-            SelectedTrack = null;
-
-            // Disconnect NTRIP if connected
-            if (_ntripService.IsConnected)
-            {
-                await _ntripService.DisconnectAsync();
-            }
-
-            // Clear field service state
-            _fieldService.SetActiveField(null);
-
-            StatusMessage = "Field closed";
-        });
-
-        DriveInCommand = new RelayCommand(() =>
-        {
-            // Start a new field at current GPS position
-            if (Latitude != 0 && Longitude != 0)
-            {
-                StatusMessage = "Drive-in field started";
-            }
-        });
-
-        ResumeFieldCommand = new RelayCommand(() =>
-        {
-            var lastField = _settingsService.Settings.LastOpenedField;
-            if (string.IsNullOrEmpty(lastField))
-            {
-                StatusMessage = "No previous field to resume";
-                return;
-            }
-
-            // Get fields directory from settings (same pattern as ConfirmFieldSelectionDialogCommand)
-            var fieldsDir = _settingsService.Settings.FieldsDirectory;
-            if (string.IsNullOrEmpty(fieldsDir))
-            {
-                fieldsDir = Path.Combine(
-                    Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments),
-                    "AgValoniaGPS", "Fields");
-            }
-
-            var fieldPath = Path.Combine(fieldsDir, lastField);
-
-            if (!Directory.Exists(fieldPath))
-            {
-                StatusMessage = $"Field not found: {lastField}";
-                return;
-            }
-
-            try
-            {
-                FieldsRootDirectory = fieldsDir;
-                CurrentFieldName = lastField;
-                IsFieldOpen = true;
-
-                // Load field origin from Field.txt (for coordinate conversions)
-                try
-                {
-                    var fieldInfo = _fieldPlaneFileService.LoadField(fieldPath);
-                    if (fieldInfo.Origin != null)
-                    {
-                        _fieldOriginLatitude = fieldInfo.Origin.Latitude;
-                        _fieldOriginLongitude = fieldInfo.Origin.Longitude;
-                    }
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogDebug($"[Field] Could not load Field.txt origin: {ex.Message}");
-                }
-
-                // Load boundary from field (same pattern as ConfirmFieldSelectionDialogCommand)
-                var boundary = _boundaryFileService.LoadBoundary(fieldPath);
-                if (boundary != null)
-                {
-                    SetCurrentBoundary(boundary);
-                    CenterMapOnBoundary(boundary);
-
-                    // Debug: show boundary extents
-                    if (boundary.OuterBoundary?.Points.Count > 0)
-                    {
-                        var pts = boundary.OuterBoundary.Points;
-                        double minE = pts.Min(p => p.Easting);
-                        double maxE = pts.Max(p => p.Easting);
-                        double minN = pts.Min(p => p.Northing);
-                        double maxN = pts.Max(p => p.Northing);
-                        _logger.LogDebug($"[Boundary] Extents (local): E({minE:F1} to {maxE:F1}), N({minN:F1} to {maxN:F1})");
-                    }
-                }
-
-                // Load background image from field
-                LoadBackgroundImage(fieldPath, boundary);
-
-                // Set the active field so headland and other field-specific data loads
-                var field = new Field
-                {
-                    Name = lastField,
-                    DirectoryPath = fieldPath,
-                    Boundary = boundary
-                };
-                _fieldService.SetActiveField(field);
-
-                // Handle NTRIP profile for this field
-                _ = HandleNtripProfileForFieldAsync(lastField);
-
-                IsJobMenuPanelVisible = false;
-                StatusMessage = $"Resumed field: {lastField}";
-            }
-            catch (Exception ex)
-            {
-                StatusMessage = $"Failed to load field: {ex.Message}";
-            }
-        });
-
-        // Map Commands
-        Toggle3DModeCommand = new RelayCommand(() =>
-        {
-            _mapService.Toggle3DMode();
-            Is2DMode = !_mapService.Is3DMode;
-        });
-
-        ZoomInCommand = new RelayCommand(() =>
-        {
-            _mapService.Zoom(1.2);
-            ZoomInRequested?.Invoke();
-        });
-
-        ZoomOutCommand = new RelayCommand(() =>
-        {
-            _mapService.Zoom(0.8);
-            ZoomOutRequested?.Invoke();
-        });
-
-        // Boundary Recording Commands
-        ToggleBoundaryPanelCommand = new RelayCommand(() =>
-        {
-            IsBoundaryPanelVisible = !IsBoundaryPanelVisible;
-        });
-
-        StartBoundaryRecordingCommand = new RelayCommand(() =>
-        {
-            _boundaryRecordingService.StartRecording(BoundaryType.Outer);
-            StatusMessage = "Boundary recording started";
-        });
-
-        PauseBoundaryRecordingCommand = new RelayCommand(() =>
-        {
-            _boundaryRecordingService.PauseRecording();
-            IsBoundaryRecording = false;
-            StatusMessage = "Boundary recording paused";
-        });
-
-        StopBoundaryRecordingCommand = new RelayCommand(() =>
-        {
-            var polygon = _boundaryRecordingService.StopRecording();
-
-            if (polygon != null && polygon.Points.Count >= 3)
-            {
-                // Save to current field
-                if (!string.IsNullOrEmpty(CurrentFieldName))
-                {
-                    var fieldPath = Path.Combine(_settingsService.Settings.FieldsDirectory, CurrentFieldName);
-                    var boundary = _boundaryFileService.LoadBoundary(fieldPath) ?? new Boundary();
-                    boundary.OuterBoundary = polygon;
-                    _boundaryFileService.SaveBoundary(boundary, fieldPath);
-                    SetCurrentBoundary(boundary);
-                    RefreshBoundaryList();
-                    StatusMessage = $"Boundary saved with {polygon.Points.Count} points, Area: {polygon.AreaHectares:F2} Ha";
-                }
-                else
-                {
-                    StatusMessage = "Cannot save boundary - no field is open";
-                }
-            }
-            else
-            {
-                StatusMessage = "Boundary not saved - need at least 3 points";
-            }
-
-            // Hide the player panel
-            IsBoundaryPlayerPanelVisible = false;
-            IsBoundaryRecording = false;
-        });
-
-        ToggleRecordingCommand = new RelayCommand(() =>
-        {
-            if (IsBoundaryRecording)
-            {
-                _boundaryRecordingService.PauseRecording();
-                IsBoundaryRecording = false;
-                StatusMessage = "Recording paused";
-            }
-            else
-            {
-                _boundaryRecordingService.ResumeRecording();
-                IsBoundaryRecording = true;
-                StatusMessage = "Recording boundary - drive around the perimeter";
-            }
-        });
-
-        UndoBoundaryPointCommand = new RelayCommand(() =>
-        {
-            _boundaryRecordingService.RemoveLastPoint();
-        });
-
-        ClearBoundaryCommand = new RelayCommand(() =>
-        {
-            _boundaryRecordingService.ClearPoints();
-            StatusMessage = "Boundary cleared";
-        });
-
-        AddBoundaryPointCommand = new RelayCommand(() =>
-        {
-            double headingRadians = Heading * Math.PI / 180.0;
-            var (offsetEasting, offsetNorthing) = CalculateOffsetPosition(Easting, Northing, headingRadians);
-            _boundaryRecordingService.AddPointManual(offsetEasting, offsetNorthing, headingRadians);
-            StatusMessage = $"Point added ({_boundaryRecordingService.PointCount} total)";
-        });
-
-        ToggleBoundaryLeftRightCommand = new RelayCommand(() =>
-        {
-            IsDrawRightSide = !IsDrawRightSide;
-        });
-
-        ToggleBoundaryAntennaToolCommand = new RelayCommand(() =>
-        {
-            IsDrawAtPivot = !IsDrawAtPivot;
-        });
-
-        ShowBoundaryOffsetDialogCommand = new RelayCommand(() =>
-        {
-            // Show numeric input dialog for boundary offset
-            NumericInputDialogTitle = "Boundary Offset (cm)";
-            NumericInputDialogValue = (decimal)BoundaryOffset;
-            NumericInputDialogDisplayText = BoundaryOffset.ToString("F0");
-            NumericInputDialogIntegerOnly = true;
-            NumericInputDialogAllowNegative = false;
-            _numericInputDialogCallback = (value) =>
-            {
-                BoundaryOffset = value;
-                StatusMessage = $"Boundary offset set to {BoundaryOffset:F0} cm";
-            };
-            State.UI.ShowDialog(DialogType.NumericInput);
-        });
-
-        CancelNumericInputDialogCommand = new RelayCommand(() =>
-        {
-            State.UI.CloseDialog();
-            _numericInputDialogCallback = null;
-        });
-
-        ConfirmNumericInputDialogCommand = new RelayCommand(() =>
-        {
-            if (NumericInputDialogValue.HasValue && _numericInputDialogCallback != null)
-            {
-                _numericInputDialogCallback((double)NumericInputDialogValue.Value);
-            }
-            State.UI.CloseDialog();
-            _numericInputDialogCallback = null;
-        });
-
-        DeleteBoundaryCommand = new RelayCommand(DeleteSelectedBoundary);
-
-        ImportKmlBoundaryCommand = new AsyncRelayCommand(async () =>
-        {
-            // Must have a field open
-            if (!IsFieldOpen || string.IsNullOrEmpty(CurrentFieldName))
-            {
-                StatusMessage = "Open a field first before importing a boundary";
-                return;
-            }
-
-            var fieldPath = Path.Combine(_settingsService.Settings.FieldsDirectory, CurrentFieldName);
-            var result = await _dialogService.ShowKmlImportDialogAsync(_settingsService.Settings.FieldsDirectory, fieldPath);
-
-            if (result != null && result.BoundaryPoints.Count > 0)
-            {
-                try
-                {
-                    // Load existing boundary or create new one
-                    var boundary = _boundaryFileService.LoadBoundary(fieldPath) ?? new Boundary();
-
-                    // Convert WGS84 boundary points to local coordinates
-                    var origin = new Wgs84(result.CenterLatitude, result.CenterLongitude);
-                    var sharedProps = new SharedFieldProperties();
-                    var localPlane = new LocalPlane(origin, sharedProps);
-
-                    var outerPolygon = new BoundaryPolygon();
-
-                    foreach (var (lat, lon) in result.BoundaryPoints)
-                    {
-                        var wgs84 = new Wgs84(lat, lon);
-                        var geoCoord = localPlane.ConvertWgs84ToGeoCoord(wgs84);
-                        outerPolygon.Points.Add(new BoundaryPoint(geoCoord.Easting, geoCoord.Northing, 0));
-                    }
-
-                    boundary.OuterBoundary = outerPolygon;
-
-                    // Save boundary
-                    _boundaryFileService.SaveBoundary(boundary, fieldPath);
-
-                    // Update map
-                    SetCurrentBoundary(boundary);
-
-                    // Refresh the boundary list
-                    RefreshBoundaryList();
-
-                    StatusMessage = $"Boundary imported from KML ({outerPolygon.Points.Count} points)";
-                }
-                catch (Exception ex)
-                {
-                    StatusMessage = $"Error importing KML boundary: {ex.Message}";
-                }
-            }
-        });
-
-        DrawMapBoundaryCommand = new RelayCommand(() =>
-        {
-            // Must have a field open
-            if (!IsFieldOpen || string.IsNullOrEmpty(CurrentFieldName))
-            {
-                StatusMessage = "Open a field first to add boundary";
-                return;
-            }
-
-            // Use the shared panel-based dialog (works on iOS and Desktop)
-            ShowBoundaryMapDialogCommand?.Execute(null);
-        });
-
-        // Keep Desktop-only async version for IDialogService integration
-        DrawMapBoundaryDesktopCommand = new AsyncRelayCommand(async () =>
-        {
-            // Must have a field open
-            if (!IsFieldOpen || string.IsNullOrEmpty(CurrentFieldName))
-            {
-                StatusMessage = "Open a field first to add boundary";
-                return;
-            }
-
-            var result = await _dialogService.ShowMapBoundaryDialogAsync(Latitude, Longitude);
-
-            if (result != null && (result.BoundaryPoints.Count >= 3 || result.HasBackgroundImage))
-            {
-                try
-                {
-                    var fieldPath = Path.Combine(_settingsService.Settings.FieldsDirectory, CurrentFieldName);
-                    LocalPlane? localPlane = null;
-
-                    if (result.BoundaryPoints.Count >= 3)
-                    {
-                        // Calculate center of boundary points
-                        double sumLat = 0, sumLon = 0;
-                        foreach (var point in result.BoundaryPoints)
-                        {
-                            sumLat += point.Latitude;
-                            sumLon += point.Longitude;
-                        }
-                        double centerLat = sumLat / result.BoundaryPoints.Count;
-                        double centerLon = sumLon / result.BoundaryPoints.Count;
-
-                        var origin = new Wgs84(centerLat, centerLon);
-                        var sharedProps = new SharedFieldProperties();
-                        localPlane = new LocalPlane(origin, sharedProps);
-                    }
-                    else if (result.HasBackgroundImage)
-                    {
-                        // Use background image center as origin
-                        double centerLat = (result.NorthWestLat + result.SouthEastLat) / 2;
-                        double centerLon = (result.NorthWestLon + result.SouthEastLon) / 2;
-
-                        var origin = new Wgs84(centerLat, centerLon);
-                        var sharedProps = new SharedFieldProperties();
-                        localPlane = new LocalPlane(origin, sharedProps);
-                    }
-
-                    // Process boundary points if present
-                    if (result.BoundaryPoints.Count >= 3 && localPlane != null)
-                    {
-                        var boundary = new Boundary();
-                        var outerPolygon = new BoundaryPolygon();
-
-                        foreach (var point in result.BoundaryPoints)
-                        {
-                            var wgs84 = new Wgs84(point.Latitude, point.Longitude);
-                            var geoCoord = localPlane.ConvertWgs84ToGeoCoord(wgs84);
-                            outerPolygon.Points.Add(new BoundaryPoint(geoCoord.Easting, geoCoord.Northing, 0));
-                        }
-
-                        boundary.OuterBoundary = outerPolygon;
-
-                        // Save boundary
-                        _boundaryFileService.SaveBoundary(boundary, fieldPath);
-
-                        // Update Field.txt with the origin used for this boundary
-                        double originLat = localPlane.Origin.Latitude;
-                        double originLon = localPlane.Origin.Longitude;
-                        _fieldOriginLatitude = originLat;
-                        _fieldOriginLongitude = originLon;
-                        try
-                        {
-                            var fieldInfo = _fieldPlaneFileService.LoadField(fieldPath);
-                            fieldInfo.Origin = new Position { Latitude = originLat, Longitude = originLon };
-                            _fieldPlaneFileService.SaveField(fieldInfo, fieldPath);
-                            _logger.LogDebug($"[MapBoundary] Updated Field.txt origin to ({originLat:F8}, {originLon:F8})");
-                        }
-                        catch (Exception ex)
-                        {
-                            _logger.LogDebug($"[MapBoundary] Could not update Field.txt: {ex.Message}");
-                        }
-
-                        // Update map
-                        SetCurrentBoundary(boundary);
-                        CenterMapOnBoundary(boundary);
-
-                        // Refresh the boundary list
-                        RefreshBoundaryList();
-                    }
-
-                    // Process background image if present
-                    if (result.HasBackgroundImage && !string.IsNullOrEmpty(result.BackgroundImagePath))
-                    {
-                        SaveBackgroundImage(result.BackgroundImagePath, fieldPath,
-                            result.NorthWestLat, result.NorthWestLon,
-                            result.SouthEastLat, result.SouthEastLon);
-                    }
-
-                    // Build status message
-                    var msgParts = new System.Collections.Generic.List<string>();
-                    if (result.BoundaryPoints.Count >= 3)
-                        msgParts.Add($"boundary ({result.BoundaryPoints.Count} pts)");
-                    if (result.HasBackgroundImage)
-                        msgParts.Add("background image");
-
-                    StatusMessage = $"Imported from satellite map: {string.Join(" + ", msgParts)}";
-                }
-                catch (Exception ex)
-                {
-                    StatusMessage = $"Error importing: {ex.Message}";
-                }
-            }
-        });
-
-        BuildFromTracksCommand = new RelayCommand(() =>
-        {
-            StatusMessage = "Build boundary from tracks not yet implemented";
-        });
-
-        DriveAroundFieldCommand = new RelayCommand(() =>
-        {
-            // Must have a field open
-            if (!IsFieldOpen || string.IsNullOrEmpty(CurrentFieldName))
-            {
-                StatusMessage = "Open a field first before recording a boundary";
-                return;
-            }
-
-            // Hide boundary panel, show the player panel
-            IsBoundaryPanelVisible = false;
-            IsBoundaryPlayerPanelVisible = true;
-
-            // Initialize recording service for a new boundary (paused state)
-            _boundaryRecordingService.StartRecording(BoundaryType.Outer);
-            _boundaryRecordingService.PauseRecording();
-
-            StatusMessage = "Drive around the field boundary. Click Record to start.";
-        });
-    }
 
     private void CenterMapOnBoundary(Boundary boundary)
     {
